Moved console to it's own folder, it doesn't work though so everything needs to be reconstructed.
