+++ src/App.jsx
@@ -3,6 +3,7 @@
  */
 import React from 'react';
 import Home from './app/screens/Home';
+import Terminal from './app/screens/Terminal/terminal.jsx'
 
 /**
  * Initialize firebase and go to AuthNavigator
@@ -10,7 +11,7 @@ import Home from './app/screens/Home';
 function App() {
   return (
     <div className="App">
-      <Home />
+      <Terminal />
     </div>
   );
 }
