+++ src/app/screens/Terminal/console.css
@@ -1,0 +1,20 @@
+.ConsoleContainer {
+    display: flex;
+    flex-direction: column;
+
+    transition: width 0.3s;
+
+    background-color: #101010;
+    font-palette: light;
+}
+
+.ConsoleSelector {
+    display: flex;
+    flex-direction: row;
+
+    position: relative;
+
+    width: 100%;
+    height: 5%;
+    background-color: #202020;
+}
