+++ src/app/screens/Terminal/terminal.css
@@ -29,27 +29,6 @@
     background-color: #555555;
 }
 
-.ConsoleContainer {
-    display: flex;
-    flex-direction: column;
-
-    transition: width 0.3s;
-
-    background-color: #101010;
-    font-palette: light;
-}
-
-.ConsoleSelector {
-    display: flex;
-    flex-direction: row;
-
-    position: relative;
-
-    width: 100%;
-    height: 5%;
-    background-color: #202020;
-}
-
 .ChatToggleButton {
     position: absolute;
     left: 10px;
